modefy and modulize go_serial
