+++ go_serial/loop_print/main.go
@@ -1,72 +1,23 @@
 package main
 
 import (
-	"bufio"
 	"fmt"
+	"go_serial/lib"
 	"log"
-	"os"
 	"time"
-
-	"go.bug.st/serial"
 )
 
-func ReadProgram(filename string) string {
-	file, err := os.Open(filename)
-	if err != nil {
-		panic(err)
-	}
-	defer file.Close()
-
-	scanner := bufio.NewScanner(file)
-	var programLines []string
-	for scanner.Scan() {
-		programLines = append(programLines, scanner.Text()+"\r\n") // ここを変更
-	}
-	if err := scanner.Err(); err != nil {
-		panic(err)
-	}
-
-	program := []byte{}
-	for _, line := range programLines {
-		program = append(program, []byte(line)...)
-	}
-
-	return string(program)
-}
-
-func programExecuteLoop(program string, port serial.Port) {
-
-	//execute program
-	fmt.Println("serial connected")
-	port.Write([]byte("edit 1 \r"))
-	time.Sleep(100 * time.Millisecond)
-	n, err := port.Write([]byte(program))
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	fmt.Printf("Sent %v bytes \n", n)
-	time.Sleep(100 * time.Millisecond)
-	port.Write([]byte("edit 0 \r"))
-	time.Sleep(100 * time.Millisecond)
-
-	port.Write([]byte("run \r"))
-
-}
 func main() {
-	mode := &serial.Mode{
-		BaudRate: 115200,
-	}
-	port, err := serial.Open("/dev/ttyUSB0", mode)
+	err, port := lib.OpenPort()
 	if err != nil {
 		log.Fatal(err)
 	}
-	defer port.Close()
-	filename := "src/print_loop.txt"
-	program := ReadProgram(filename)
 
-	go programExecuteLoop(program, port)
+	defer port.Close()
+	filename := "../basic_src/print_loop.txt"
+	program := lib.ReadProgram(filename)
 	time.Sleep(time.Second)
+	go lib.ProgramExecute(program, port)
 
 	buff := make([]byte, 300)
 	for {
@@ -80,50 +31,3 @@ func main() {
 	}
 
 }
-
-// func programExecute(program string) {
-// 	mode := &serial.Mode{
-// 		BaudRate: 115200,
-// 	}
-// 	port, err := serial.Open("/dev/ttyUSB0", mode)
-// 	if err != nil {
-// 		log.Fatal(err)
-// 	}
-// 	defer port.Close()
-
-// 	//execute program
-// 	fmt.Println("serial connected")
-// 	port.Write([]byte("edit 1 \r"))
-// 	time.Sleep(100 * time.Millisecond)
-// 	n, err := port.Write([]byte(program))
-// 	if err != nil {
-// 		log.Fatal(err)
-// 	}
-
-// 	fmt.Printf("Sent %v bytes \n", n)
-// 	time.Sleep(100 * time.Millisecond)
-// 	port.Write([]byte("edit 0 \r"))
-// 	time.Sleep(100 * time.Millisecond)
-// 	buff := make([]byte, 1000)
-// 	eofFound := false
-
-// 	port.Write([]byte("run \r"))
-
-// 	for !eofFound {
-// 		// Reads up to 100 bytes
-// 		n, err := port.Read(buff)
-// 		if err != nil {
-// 			log.Fatal(err)
-// 		}
-
-// 		receivedData := string(buff[:n])
-// 		fmt.Printf("%s", receivedData)
-
-// 		if strings.Contains(receivedData, "EOF") {
-// 			fmt.Println("\nEOF")
-// 			eofFound = true
-// 		}
-// 		// log.Fatal("can't recive valid data")
-// 	}
-
-// }
